A basic system test.
Added the RAILS_SYSTEM_TESTING_SCREENSHOT=simple environment
variable to show only the path to the screenshot on failure.
Fiddled with the prototype to make it still work (I think)
with a User model. To make the prototype work, you have to do:
`rails db:fixtures:load` in the development environment. This
is just a temporary step.