Recursive queries in database
I tried doing recursive queries in the database. I got the queries
to work. An example is in ci.rb. Fitting them into an association
was challenging because the association adds a where for the
foreign key that I couldn't get rid of, which restricted the rows
returned from the recursive query to the immediate ancestor or
descendant. In addition, at least as of five years ago, Postgres
wouldn't optimize the query through to the original query, so we
might have had performance issues anyway. Worth pursuing further,
but enough time wasted now. Will tag this so I can find it again.